Simplified blockly "Functions" category.
I removed the "function" and the "if-return" blocks. I added two "return"
blocks (with value and without value) and I plan to use the "procedure"
block to represent both procedures and functions. Using the new "return"
blocks I expect to be able to distinguis between them and generate the
correct code.